requirements.txt
bidict==0.22.1
click==8.1.3
dnspython==1.16.0
eventlet==0.30.2
Flask==2.2.2
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
six==1.16.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
zipp==3.11.0
